sim/avr_usb: move types to separate file
